GTK: Rename 'druid_gtk' platform to 'gtk'
